user scale as part of SHIFT + SCALE cycling
- Song::userScale stores user scale (there can be only one)

- scaleId() -> NUM_PRESET_SCALES for user scales

- scale name is always fetched with getScaleName()

- OTHER_SCALE and CUSTOM_SCALE_WITH_MORE_THAN_7_NOTES both replaced
  with USER_SCALE, or CANT_CHANGE_SCALE as appropriate.

- propagating the correct type for scale id / index a bit
  furhter, still remains as int32_t farther out.

+++ src/deluge/model/clip/instrument_clip_minder.h
@@ -43,7 +43,7 @@ public:
 	void drawActualNoteCode(int16_t noteCode);
 	void cycleThroughScales();
 	bool setScale(int32_t newScale);
-	void displayScaleName(int32_t scale);
+	void displayScaleName(uint8_t scale);
 	void displayCurrentScaleName();
 	void selectEncoderAction(int32_t offset);
 	static void drawMIDIControlNumber(int32_t controlNumber, bool automationExists);
